The most significant changes involve the addition of two new files, `ActivityExtensions.cs` and `MigrateDbContextExtensions.cs`, which contain static classes for handling exception tags and database migrations respectively.
In `ActivityExtensions.cs`, a new static class `ActivityExtensions` has been added. This class contains a method `SetExceptionTags` which is an extension method for the `Activity` class. This method is used to add tags to an `Activity` instance related to an exception such as the exception message, stack trace, and type. It also sets the status of the `Activity` to `Error`.

In `MigrateDbContextExtensions.cs`, a new static class `MigrateDbContextExtensions` has been added. This class contains several methods related to database migrations. The `AddMigration` methods are extension methods for `IServiceCollection` that add migration services for a specific `DbContext`. The `MigrateDbContextAsync` method is an extension method for `IServiceProvider` that migrates a specific `DbContext` and invokes a seeder function. The `InvokeSeeder` method and `MigrationHostedService` class are used to migrate a specific `DbContext` and invoke a seeder function.

A new public interface `IDbSeeder` has been added which contains a method `SeedAsync` for seeding a specific `DbContext`.

Changes:
1. Addition of `ActivityExtensions.cs` file with `ActivityExtensions` static class and `SetExceptionTags` method.
2. Addition of `MigrateDbContextExtensions.cs` file with `MigrateDbContextExtensions` static class, `AddMigration` methods, `MigrateDbContextAsync` method, `InvokeSeeder` method, and `MigrationHostedService` class.
3. Addition of `IDbSeeder` interface with `SeedAsync` method.